Added Activity draw
Added Activity draw to Original Project and setup new project because I needed to add fragments for the video and was having a hard time doing that without just using a template. I tried to add the template to the original project but is was more effort than it would have been worth.